pyproject.toml
[project]
name = "modelplayground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.23",
    "flask>=3.1.1",
    "google-genai>=1.16.1",
    "python-dotenv>=1.1.0",
]
